Start extracting medallia code
