+++ beater/journalbeat.go
@@ -22,8 +22,6 @@ import (
 	"strconv"
 	"time"
 
-	"hash/fnv"
-
 	"github.com/coreos/go-systemd/sdjournal"
 	"github.com/elastic/beats/libbeat/beat"
 	"github.com/elastic/beats/libbeat/common"
@@ -37,67 +35,18 @@ import (
 	"github.com/wavefronthq/go-metrics-wavefront"
 )
 
-type LogBuffer struct {
-	time     time.Time
-	logEvent common.MapStr
-	logType  string
-}
-
-func hash(s string) int {
-	h := fnv.New32a()
-	h.Write([]byte(s))
-	return int(h.Sum32())
-}
-
-func getPartition(lb *LogBuffer, numPartitions int) int {
-	partition := 0
-	if tag, ok := lb.logEvent["container_tag"]; ok {
-		// same container - same instance
-		// Assuming equal config - if container moves, it should still
-		// end up at same logstash instance
-		partition = hash(tag.(string)) % numPartitions
-	} else if buftype, ok := lb.logEvent["logBufferingType"]; ok {
-		// journalbeat does re-assembly based on logBufferingType
-		partition = hash(buftype.(string)) % numPartitions
-	} else if eventtype, ok := lb.logEvent["type"]; ok {
-		partition = hash(eventtype.(string)) % numPartitions
-	}
-	return partition
-}
-
-const (
-	metricPrefix string = "logging.journalbeat"
-	//These are the fields for the container logs.
-	containerTagField string = "CONTAINER_TAG"
-	containerIdField  string = "CONTAINER_ID"
-
-	//These are the fields for the host process logs.
-	tagField     string = "SYSLOG_IDENTIFIER"
-	processField string = "_PID"
-
-	//Common fields for both container and host process logs.
-	hostNameField  string = "_HOST_NAME"
-	messageField   string = "MESSAGE"
-	timestampField string = "_SOURCE_REALTIME_TIMESTAMP"
-	priorityField  string = "PRIORITY"
-
-	channelSize   int   = 1000
-	microseconds  int64 = 1000000
-	microsToNanos int64 = 1000
-)
-
 // Journalbeat is the main Journalbeat struct
 type Journalbeat struct {
-	done                 chan struct{}
-	config               config.Config
-	client               publisher.Client
-	logstashClients      []publisher.Client
-	numLogstashAvailable int //corresponds to the number of downstream logstash aggregators available at startup.
+	done   chan struct{}
+	config config.Config
+	client publisher.Client
 
 	journal *sdjournal.Journal
 
 	cursorChan chan string
 
+	logstashClients                 []publisher.Client
+	numLogstashAvailable            int //corresponds to the number of downstream logstash aggregators available at startup.
 	journalTypeOutstandingLogBuffer map[string]*LogBuffer
 	incomingLogMessages             chan common.MapStr
 
@@ -214,99 +163,6 @@ func New(b *beat.Beat, cfg *common.Config) (beat.Beater, error) {
 	return jb, nil
 }
 
-func (jb *Journalbeat) flushStaleLogMessages() {
-	for logType, logBuffer := range jb.journalTypeOutstandingLogBuffer {
-		if time.Now().Sub(logBuffer.time).Seconds() >= jb.config.FlushLogInterval.Seconds() {
-			//this message has been sitting in our buffer for more than 30 seconds time to flush it.
-			partition := getPartition(logBuffer, jb.numLogstashAvailable)
-			jb.logstashClients[partition].PublishEvent(logBuffer.logEvent, publisher.Guaranteed)
-			delete(jb.journalTypeOutstandingLogBuffer, logType)
-			jb.cursorChan <- logBuffer.logEvent["cursor"].(string)
-		}
-	}
-}
-
-func (jb *Journalbeat) flushOrBufferLogs(event common.MapStr) {
-	//check if it starts with space or tab
-	newLogMessage := event["message"].(string)
-	logType := event["logBufferingType"].(string)
-
-	if newLogMessage != "" && (newLogMessage[0] == ' ' || newLogMessage[0] == '\t') {
-		//this is a continuation of previous line
-		if oldLog, found := jb.journalTypeOutstandingLogBuffer[logType]; found {
-			jb.journalTypeOutstandingLogBuffer[logType].logEvent["message"] =
-				oldLog.logEvent["message"].(string) + "\n" + newLogMessage
-		} else {
-			jb.journalTypeOutstandingLogBuffer[logType] = &LogBuffer{
-				time:     time.Now(),
-				logType:  event["logBufferingType"].(string),
-				logEvent: event,
-			}
-		}
-		jb.journalTypeOutstandingLogBuffer[logType].time = time.Now()
-	} else {
-		oldLogBuffer, found := jb.journalTypeOutstandingLogBuffer[logType]
-		jb.journalTypeOutstandingLogBuffer[logType] = &LogBuffer{
-			time:     time.Now(),
-			logType:  event["logBufferingType"].(string),
-			logEvent: event,
-		}
-		if found {
-			//flush the older logs to async.
-			partition := getPartition(oldLogBuffer, jb.numLogstashAvailable)
-			jb.logstashClients[partition].PublishEvent(oldLogBuffer.logEvent, publisher.Guaranteed)
-			//update stats if enabled
-			if jb.config.MetricsEnabled {
-				jb.logMessagesPublished.Inc(1)
-				jb.logMessageDelay.Update(time.Now().Unix() - (event["utcTimestamp"].(int64) / microseconds))
-			}
-		}
-	}
-}
-
-//TODO optimize this later but for now walkthru all the different types. Use priority queue/multiple threads if needed.
-func (jb *Journalbeat) logProcessor() {
-	logp.Info("Started the thread which consumes log messages and publishes it")
-	tickChan := time.NewTicker(jb.config.FlushLogInterval)
-	for {
-		select {
-		case <-tickChan.C:
-			//here we need to walk thru all the map entries and flush out the ones
-			//which have been sitting there for some time.
-			jb.flushStaleLogMessages()
-
-		case channelEvent := <-jb.incomingLogMessages:
-			jb.flushOrBufferLogs(channelEvent)
-		}
-	}
-}
-
-// "circular shift" a config list
-func shiftlist(cfg *common.Config, target *common.Config, key string, shift int) error {
-	count, err := cfg.CountField(key)
-	if err != nil {
-		return err
-	}
-	offset := 0
-	for n := shift; n < count; n++ {
-		item, err := cfg.String(key, n)
-		if err != nil {
-			return err
-		}
-		target.SetString(key, offset, item)
-		offset++
-	}
-	for n := 0; n < shift; n++ {
-		item, err := cfg.String(key, n)
-		if err != nil {
-			return err
-		}
-		target.SetString(key, offset, item)
-		offset++
-	}
-	return nil
-}
-
 // Run is the main event loop: read from journald and pass it to Publish
 func (jb *Journalbeat) Run(b *beat.Beat) error {
 	logp.Info("Journalbeat is running!")
@@ -356,8 +212,8 @@ func (jb *Journalbeat) Run(b *beat.Beat) error {
 				jb.config.MetricsInterval, // interval
 				jb.config.InfluxDBURL,     // the InfluxDB url
 				jb.config.InfluxDatabase,  // your InfluxDB database
-				"", // your InfluxDB user
-				"", // your InfluxDB password
+				"",              // your InfluxDB user
+				"",              // your InfluxDB password
 			)
 		}
 	}
@@ -384,27 +240,27 @@ func (jb *Journalbeat) Run(b *beat.Beat) error {
 	}
 
 	for i := 0; i < jb.numLogstashAvailable; i++ {
-		processors, err := processors.New(b.Config.Processors)
+		newProcessors, err := processors.New(b.Config.Processors)
 		if err != nil {
 			return fmt.Errorf("error initializing processors: %v", err)
 		}
 
-		//override the hosts to pick one of the entries from the original hosts configuration.
-		newoutput, err := common.NewConfigFrom(b.Config.Output["logstash"])
+		// override the hosts to pick one of the entries from the original hosts configuration.
+		newOutput, err := common.NewConfigFrom(b.Config.Output["logstash"])
 		if err != nil {
 			return fmt.Errorf("Failed to clone output config: %v", err)
 		}
-		err = shiftlist(b.Config.Output["logstash"], newoutput, "hosts", i)
+		err = shiftlist(b.Config.Output["logstash"], newOutput, "hosts", i)
 		if err != nil {
 			return fmt.Errorf("Failed to shift list %v", err)
 		}
 
-		publisher, err := publisher.New(b.Name, b.Version, map[string]*common.Config{"logstash": newoutput}, b.Config.Shipper, processors)
+		newPublisher, err := publisher.New(b.Name, b.Version, map[string]*common.Config{"logstash": newOutput}, b.Config.Shipper, newProcessors)
 		if err != nil {
 			return fmt.Errorf("error initializing publisher: %v", err)
 		}
 
-		jb.logstashClients = append(jb.logstashClients, publisher.Connect())
+		jb.logstashClients = append(jb.logstashClients, newPublisher.Connect())
 	}
 
 	commonFields := []string{hostNameField, messageField, priorityField}
@@ -458,4 +314,5 @@ func (jb *Journalbeat) Stop() {
 	for i := 0; i < jb.numLogstashAvailable; i++ {
 		jb.logstashClients[i].Close()
 	}
+	logp.Info("Journalbeat stopped")
 }
